finished

+++ application/controllers/Blog.php
@@ -20,6 +20,8 @@ class Blog extends CI_Controller
 
 	public function article($id = NULL)
 	{
+
+
 		if (!is_numeric($id)) {
 			redirect('blog/index');
 		}
@@ -27,6 +29,9 @@ class Blog extends CI_Controller
 		$this->load->model('article');
 		$this->load->model('article_status');
 		$this->article->load($id, $this->auth_user->is_connected);
+
+		$data['title'] = htmlentities($this->article->title);
+		$data['script'] = '<script src="' . base_url('js/article.js') . '"></script>';
 
 		if ($this->article->is_found) {
 			$data['title'] = htmlentities($this->article->title);
@@ -95,16 +100,40 @@ class Blog extends CI_Controller
 		if (!$this->article->is_found) {
 			redirect('blog/index');
 		}
-		if ($this->input->post('confirm') === NULL) {
-			$data['action'] = "confirm";
-		} else {
-			$this->article->delete();
-			$data['action'] = "result";
-		}
-		$data['title'] = "Suppression article";
 		$this->load->helper('form');
-		$this->load->view('common/header', $data);
-		$this->load->view('blog/delete', $data);
-		$this->load->view('common/footer', $data);
+		if ($this->input->is_ajax_request()) {
+			// nous avons reçu une requête ajax
+			$this->load->view('blog/delete_confirm');
+		} else {
+			// nous avons reçu une requête classique
+			if ($this->input->post('confirm') === NULL) {
+				$data['action'] = "confirm";
+			} else {
+				$this->article->delete();
+				$data['action'] = "result";
+			}
+			$data['title'] = "Suppression article";
+			$this->load->view('common/header', $data);
+			$this->load->view('blog/delete', $data);
+			$this->load->view('common/footer', $data);
+		}
+	}
+
+	public function publication($id = NULL)
+	{
+		if (!$this->auth_user->is_connected) {
+			redirect('blog/index');
+		}
+		if (!is_numeric($id)) {
+			redirect('blog/index');
+		}
+		$this->load->model('article');
+		$this->article->load($id, TRUE);
+		if (!$this->article->is_found) {
+			redirect('blog/index');
+		}
+		$this->article->status = 'P';
+		$this->article->save();
+		redirect('blog/' . $this->article->alias . '_' . $id);
 	}
 }
